requirements.txt
discord.py
python-dotenv
disnake
discord-py-slash-command
discord-components
requests
asyncio
pynacl
discord.py[voice]
typing
utils
googletrans
youtube_dl
Pillow
youtube_dl